moving flashcards/... from /pkg to /internal
